+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.1] - 2024-10-10
+### Added
+- Added properties ordering rules for display, positioning, dimensions, and box model.
+- Added ordering for unspecified properties  with `unspecified: bottomAlphabetical`
+
 ## [1.0.0] - 2024-10-02
 ### Added
 - Initial project setup.